Two files added that are proposed final versions

+++ RMS_extra_tools/Error_Check_and_Cleanup.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
-# ErrorCheck.sh
-# Command line argument 1 ($1): Full path to ArchivedFiles directory
+# Error_Check_and_Cleanup.sh
+# This script requires a command line argument ("$1") containing 
+# the full path to the CapturedFiles directory for a given night
 
 archive_dir="$(dirname "$1")"
 data_dir="$(dirname "$archive_dir")"
@@ -10,8 +11,15 @@ night_dir="$(basename "$1")"
 station=${night_dir:0:6}
 OUTFILE=$data_dir"/"${station}_"fits_counts.txt"
 
+printf '\nError_Check_and_Cleanup.sh, revised 18-May, 2023, 8054 bytes,'
+printf ' was called with\nArg (directory) = %s \n' "$1"
+printf 'This script writes results to %s \n' $OUTFILE
+printf ' and can delete older files to make room for more capture directories\n\n'
+
+Cleanup=1	# set to 0 to skip data cleanup at end of script
+
 if [ ! -f "$OUTFILE" ]; then
-    # file does not exist, so write top line:
+    # file does not exist yet, so write top line of file:
     printf "Directory Name         # fits_files  # detections  Other Issues" > "$OUTFILE"
 fi
 
@@ -20,11 +28,6 @@ result=0
 short_fall=0
 secs_missed=0
 min_missed=0
-
-printf '\nErrorCheck.sh, revised 17-May, 2023, byte count 7714, '
-printf 'was called with\nArg (directory) = %s \n' "$1"
-printf 'ErrorCheck writes results to %s \n' $OUTFILE
-printf ' and can optionally clean up older files to keep older capture directories\n\n'
 
 echo data_dir:    $data_dir
 echo archive_dir: $archive_dir
@@ -58,8 +61,7 @@ if [[ $night_dir = '' || ! -d "${archive_dir}"/$night_dir ]] ; then
 fi
 
 # ____________________
-# Get capture length in seconds using newest log file
-# Find the latest log file in the log directory
+# Calculate capture length in seconds using newest log file
 capture_file=$(ls -Art $HOME/RMS_data/logs/log_*.log | tail -n 1)
 echo Checking log file: $capture_file for capture duration
 
@@ -73,28 +75,27 @@ echo hours: $hrs, seconds: $seconds, rounded off seconds: $capture_len
 
 # ____________________
 # Collect information for the output file
-# First, the number of FITS files
+# First, the number of FITS files captured
 
 fits_count=$(find "$capture_dir/$night_dir"/*.fits -type f -printf x | wc -c)
 printf '\nNumber of fits files in Capture directory: %d\n' "$fits_count"
 
-# Use the total capture time and estimate the number of fits files
+# Use the total capture time to estimate the total number of fits files
 total_fits=$(( capture_len * 100 / 1024 ))
 result="$( awk -v x="$capture_len" 'BEGIN { print ( x / 10.24 ) }' )"
 env printf "A capture lasting %d seconds, should create an estimated %d (%0.01f) fits files \n" \
 	"$capture_len"  "$total_fits" "$result"
 
-# Estimated total_fits is typically 4 more than observed, so we subtract 4 from total
-#total_fits=$(( total_fits - 4))
+# Estimated total_fits is usually 4 more than number captured, so subtract 4 from total
+total_fits=$(( total_fits - 4))
 short_fall=$(( fits_count - total_fits ))
 secs_missed=$(( short_fall * 1024 / 100 ))
 min_missed="$( awk -v x=$secs_missed 'BEGIN { print ( x / 60 ) }' )"
 env printf "We have a short fall of: %d fits, %d secs, %0.1f min\n\n" \
 	$short_fall $secs_missed $min_missed
-# Done finding the total capture time and the estimated number of fits files
 
 # ____________________
-# Find the number of detections in the stack file
+# Get the number of detections from the detected stack filename
 stack=$(ls "$archive_dir/$night_dir"/*meteors.jpg)
 # Anything such file found?
 if [[ -n $stack ]] ; then
@@ -107,9 +108,9 @@ fi
 
 # ____________________
 # Count the number of directories under CapturedFiles and ArchivedFiles
-# The variable "id_string" extracts the station name and the date (in the 
-# form yyyymmdd) from the first positional parameter (the directory # name). 
-# It uses substring syntax (take the substring starting at position # 0 and 
+# The variable "id_string" extracts the station name and the date (in the
+# form yyyymmdd) from the first positional parameter (the directory # name).
+# It uses substring syntax (take the substring starting at position # 0 and
 # extending 15 characters).
 # The variable "id_string" holds the station name and date in a pattern.
 
@@ -145,7 +146,7 @@ if [[ $num_captured_dirs -eq $num_archived_dirs ]] ; then
 	printf "%d\t%d" >> "$OUTFILE" \
 	"$fits_count" "$detected"
 
-	# echo the short fall in fits files if large enough
+	# echo any short fall in fits files
 	if [[ $short_fall -lt 0 ]] ; then
 	   printf "\t%d fits %0.1f min" >> "$OUTFILE" \
 	   "$short_fall" "$min_missed"
@@ -160,6 +161,7 @@ else
 	"$fits_count" "$detected" "$num_captured_dirs" "$num_archived_dirs"
 fi
 
+# Write a warning if there is no photometry plot
 if ! compgen -G "${archive_dir}/$night_dir/*_calib_report_photometry.png" > /dev/null ; then
    if [[ $detected -lt 3 ]] ; then
 	printf "  No Photometry, Clouded out?" >> "$OUTFILE"
@@ -169,8 +171,8 @@ if ! compgen -G "${archive_dir}/$night_dir/*_calib_report_photometry.png" > /dev
 fi
 
 # ____________________
-# Check for the TOTAL number of directories in the CapturedFiles 
-# directory, not just the number matching the date
+# Check the total number of directories in the CapturedFiles directory,
+#  not just the number matching the date, warn if there is only one found
 pushd "$capture_dir" > /dev/null
 captured=(./*"$station"*)
 total_captured=${#captured[@]}
@@ -189,49 +191,45 @@ printf "\n" >> "$OUTFILE"
 
 printf "Fits file count and number of detections saved to: %s\n\n" "$OUTFILE"
 
-# ______________________________________________________
-# The next section of the script can be used to clean up  
-# old files to free up space on the storage drive
+# ____________________________________________________________________
+# This section of the script can be used to clean up old files to free
+# up space on the storage drive for more CapturedFiles directories
+# var Cleanup is set above on line 18
 
 # set variables adirs, cdirs, and older to 0 to skip cleanups
-# delete older ArchivedFiles directories
-adirs=10
+adirs=10	# delete older ArchivedFiles directories
+cdirs=10	# delete older CapturedFiles directories
+older=10	# delete older tar.bz2 archives
+logs=21		# delete log files older than this number of days
 
-# delete older CapturedFiles directories
-cdirs=10
+if [[ $(Cleanup) -gt 0 ]]; then
+   cd "$HOME"/RMS_data/ArchivedFiles
 
-# delete older tar.bz2 archives
-older=10
+   if [[ $(adirs) -gt 0 ]]; then
+      printf "Deleting ArchivedFiles directories more than %s days old\n" "${adirs}"
+      adirs=$((adirs-1))
+      find -mtime +$adirs -type d | xargs rm -f -r
+   fi
 
-# delete older log files
-logs=21
+   if [[ ${older} -gt 0 ]]; then
+      printf "Deleting tar.bz2 files more than %s days old\n" "${older}"
+      older=$((older-1))
+      find "$HOME"/RMS_data/ArchivedFiles/*.bz2 -type f -mtime +$older -delete;
+   fi
 
-cd "$HOME"/RMS_data/ArchivedFiles
+   if [[ ${cdirs} -gt 0 ]]; then
+      cd "$HOME"/RMS_data/CapturedFiles
+      printf "Deleting CapturedFiles directories more than %s days old\n" "${cdirs}"
+      cdirs=$((cdirs-1))
+      find -mtime +$cdirs -type d | xargs rm -f -r
+   fi
 
-if [[ $(adirs) -gt 0 ]]; then
-   printf "Deleting ArchivedFiles directories more than %s days old\n" "${adirs}"
-   adirs=$((adirs-1))
-   find -mtime +$adirs -type d | xargs rm -f -r
+   if [[ ${logs} -gt 0 ]]; then
+      printf "Deleting files in RMS_data/logs more than %s days old\n" "$(logs)"
+      logs=$((logs-1))
+      find "$HOME"/RMS_data/logs/ -type f -mtime +$logs -delete;
+   fi
+
+   printf "Done deleting old data\n "
 fi
-
-if [[ ${older} -gt 0 ]]; then
-   printf "Deleting tar.bz2 files more than %s days old\n" "${older}" 
-   older=$((older-1))
-   find "$HOME"/RMS_data/ArchivedFiles/*.bz2 -type f -mtime +$older -delete;
-fi
-
-if [[ ${cdirs} -gt 0 ]]; then
-   cd "$HOME"/RMS_data/CapturedFiles
-   printf "Deleting CapturedFiles directories more than %s days old\n" "${cdirs}"
-   cdirs=$((cdirs-1))
-   find -mtime +$cdirs -type d | xargs rm -f -r 
-fi
-
-if [[ ${logs} -gt 0 ]]; then
-   printf "Deleting files in RMS_data/logs more than %s days old\n" "$(logs)"
-   logs=$((logs-1))
-   find "$HOME"/RMS_data/logs/ -type f -mtime +$logs -delete;
-fi
-
-printf "Done deleting old data\n "
-# ______________________________________________________
+# ____________________________________________________________________
